packages/cai-securegpt-client/pyproject.toml
[project]
name = "cai-securegpt-client"
version = "0.2.0"
description = ""
requires-python = ">=3.11,<4.0"
readme = "README.md"
dependencies = [ "langchain-core <3", "pydantic-core <3", "httpx_sse >=0.4", "requests_oauth2client ~=1.6.0", "python-dotenv ~=1.0.1", "langchain-openai (>=0.2.9,<0.4.0)", "mypy (>=1.16.1,<2.0.0)", "urllib3 (>=2.5.0,<3.0.0)", "requests (>=2.32.4,<3.0.0)",]
[[project.authors]]
name = "Emilio Manzano"
email = "emilio.manzano@axa.com"

[dependency-groups]
dev = [ "ruff>=0.9.6", "black>=25.1.0,<26",]
test = [ "pytest>= 8.4.1", "pytest-asyncio>=1.0.0",]

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[tool.uv]
default-groups = [ "dev", "test",]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [ "E", "W", "F", "I", "B", "C4", "N", "D",]
ignore = [ "C901", "W191", "D401", "D205", "D100", "D104",]

[tool.ruff.lint.pydocstyle]
convention = "pep257"
